feat: add service

+++ backend/bookAPI/src/main/java/com/quennel/bookapi/repository/BookRepository.java
@@ -3,5 +3,5 @@ package com.quennel.bookapi.repository;
 import com.quennel.bookapi.models.Book;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface BookRepository extends MongoRepository<Book, Integer> {
+public interface BookRepository extends MongoRepository<Book, Long> {
 }
